Cleanup the docker host in the stop hook.
The presence of the stop hook symlink being present will actually clear
up #18, but I wanted to further address this by cleaning up the host in
the event we were to co-locate services. It makes sense to clean up the
application bins.

If there is interest in cleaning up the docker images, we should make
this available as an action, and wipe all the images out with a `docker
rmi -f` - but this MP will not address that scenario