Add a call to installation-status from config-changed for first time runs.
This corrects a first time run bug, but bloats every subsequent run.
Checking on this is difficult as when checking for the presence of
docker_version, you never know if the value is actually set, which
raises a type error. We're going to have to eat a minor amount of time
to set these facts on every run to solve this unless a better method
comes along.