Refactored project files to be warehoused with tox, pytest, and et-al
The process of moving to tox is purely one of dependency isolation, and
repeatability when porting from workstaiton => CI. The virtualenv story
was strong, but tox makes most of this a no brainer.

The makefile refactors are a bit militant, but wont ask if python-tox
already exists in /usr/bin/tox (ubuntu package location) - if you're
using yet another virtualenv layer to isolate tox from your system level
dependencies, this will require you to update the Makefile, or forever
be pestered by the make targets.

THIS IS NON NEGOTIABLE

To run the new test suite tools manually:
`tox`
`tox -e lint`

The first will execute the unit tests by default

The second will execute all of the linting based commands we would expec
to see as charmers.