Swap AUFS default to True
Device mapper enabled by default causes so many issues with AWS now when
doing a vanilla deployment. Entire gigs of disk storage are consumed
pulling a simple 400MB container - with only a few layers.

This is broken behavior by default. if this causes any problems in your
setup please file a bug and lets resolve it with the AUFS provider and
fully depreciate the devicemapper config.